Move schemas, functions and layouts to a separate file
